+++ core/src/net/oschina/j2cache/CacheChannel.java
@@ -29,6 +29,11 @@ import java.util.stream.Collectors;
 public abstract class CacheChannel implements Closeable , AutoCloseable {
 
 	private final static Map<String, Object> _g_keyLocks = new ConcurrentHashMap<>();
+	private boolean support_null_obj = false;
+
+	public CacheChannel(boolean support_null_obj){
+		this.support_null_obj = support_null_obj;
+	}
 
 	/**
 	 * <p>Just for Inner Use.</p>
@@ -58,11 +63,11 @@ public abstract class CacheChannel implements Closeable , AutoCloseable {
 	public CacheObject get(String region, String key)  {
 		CacheObject obj = new CacheObject(region, key, CacheObject.LEVEL_1);
 		obj.setValue(CacheProviderHolder.getLevel1Cache(region).get(key));
-		if(obj.getValue() == null) {
+		if(obj.rawValue() == null) {
 			obj.setLevel(CacheObject.LEVEL_2);
 			obj.setValue(CacheProviderHolder.getLevel2Cache(region).get(key));
-			if(obj.getValue() != null)
-				CacheProviderHolder.getLevel1Cache(region).put(key, obj.getValue());
+			if(obj.rawValue() != null)
+				CacheProviderHolder.getLevel1Cache(region).put(key, obj.rawValue());
 		}
 		return obj;
 	}
@@ -76,11 +81,11 @@ public abstract class CacheChannel implements Closeable , AutoCloseable {
 	 */
 	public CacheObject get(String region, String key, Function<String, Object> loader) {
 		CacheObject cache = get(region, key);
-		if (cache.getValue() == null) {
+		if (cache.rawValue() == null) {
 			String lock_key = key + '@' + region;
 			synchronized (_g_keyLocks.computeIfAbsent(lock_key, v -> new Object())) {
 				cache = get(region, key);
-				if (cache.getValue() == null) {
+				if (cache.rawValue() == null) {
 					try {
 						Object obj = loader.apply(key);
 						if (obj != null) {
@@ -131,7 +136,7 @@ public abstract class CacheChannel implements Closeable , AutoCloseable {
 	 */
 	public Map<String, CacheObject> get(String region, Collection<String> keys, Function<String, Object> loader)  {
 		Map<String, CacheObject> results = get(region, keys);
-		results.entrySet().stream().filter(e -> e.getValue().getValue() == null).forEach( e -> {
+		results.entrySet().stream().filter(e -> e.getValue().rawValue() == null).forEach( e -> {
 			String lock_key = e.getKey() + '@' + region;
 			synchronized (_g_keyLocks.computeIfAbsent(lock_key, v -> new Object())) {
 				CacheObject cache = get(region, e.getKey());
@@ -175,16 +180,12 @@ public abstract class CacheChannel implements Closeable , AutoCloseable {
 	 * @param key: Cache key
 	 * @param value: Cache value
 	 */
-	public void set(String region, String key, Object value)  {
-		if(value == null)
-			evict(region, key);
-		else{
-			try {
-				CacheProviderHolder.getLevel1Cache(region).put(key, value);
-				CacheProviderHolder.getLevel2Cache(region).put(key, value);
-			} finally {
-				this.sendEvictCmd(region, key);//清除原有的一级缓存的内容
-			}
+	public void set(String region, String key, Object value) {
+		try {
+			CacheProviderHolder.getLevel1Cache(region).put(key, (value==null && support_null_obj)?new Object():value);
+			CacheProviderHolder.getLevel2Cache(region).put(key, (value==null && support_null_obj)?new Object():value);
+		} finally {
+			this.sendEvictCmd(region, key);//清除原有的一级缓存的内容
 		}
     }
 
@@ -195,19 +196,15 @@ public abstract class CacheChannel implements Closeable , AutoCloseable {
 	 * @param value Cache value
 	 * @param timeToLiveInSeconds cache expired in second
 	 */
-    public void set(String region, String key, Object value, long timeToLiveInSeconds)  {
+    public void set(String region, String key, Object value, long timeToLiveInSeconds) {
     	if(timeToLiveInSeconds <= 0)
     		set(region, key, value);
     	else {
-			if (value == null)
-				evict(region, key);
-			else {
-				try {
-					CacheProviderHolder.getLevel1Cache(region, timeToLiveInSeconds).put(key, value);
-					CacheProviderHolder.getLevel2Cache(region).put(key, value);
-				} finally {
-					this.sendEvictCmd(region, key);//清除原有的一级缓存的内容
-				}
+			try {
+				CacheProviderHolder.getLevel1Cache(region, timeToLiveInSeconds).put(key, (value==null && support_null_obj)?new Object():value);
+				CacheProviderHolder.getLevel2Cache(region).put(key, (value==null && support_null_obj)?new Object():value);
+			} finally {
+				this.sendEvictCmd(region, key);//清除原有的一级缓存的内容
 			}
 		}
 	}