+++ core/src/net/oschina/j2cache/J2Cache.java
@@ -20,8 +20,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
 
 /**
  * J2Cache 的缓存入口
@@ -33,14 +31,16 @@ public class J2Cache {
 
 	private final static String CONFIG_FILE = "/j2cache.properties";
 
+	private final static J2CacheConfig config;
 	private final static CacheChannel channel;
 	private static ClusterPolicy policy; //不同的广播策略
 
 	static {
 		try {
-			initFromConfig();
+			config = J2CacheConfig.initFromConfig(CONFIG_FILE);
+			initFromConfig(config);
 			/* 初始化缓存接口 */
-			channel = new CacheChannel(){
+			channel = new CacheChannel(config.isSupport_null_object()){
 				@Override
 				public void sendClearCmd(String region) {
 					policy.sendClearCmd(region);
@@ -75,48 +75,19 @@ public class J2Cache {
 	 * @return
 	 * @throws IOException
 	 */
-	private static void initFromConfig() throws IOException {
-		try(InputStream configStream = getConfigStream()){
-			Properties props = new Properties();
-			props.load(configStream);
-			SerializationUtils.init(props.getProperty("j2cache.serialization"));
-			//初始化两级的缓存管理
-			CacheProviderHolder.init(props, (region, key)->{
-				//当一级缓存中的对象失效时，自动清除二级缓存中的数据
-				CacheProviderHolder.getLevel2Cache(region).evict(key);
-				log.debug(String.format("Level 1 cache object expired, evict level 2 cache object [%s,%s]", region, key));
-				if(policy != null)
-					policy.sendEvictCmd(region, key);
-			});
+	private static void initFromConfig(J2CacheConfig config) {
+		SerializationUtils.init(config.getSerialization());
+		//初始化两级的缓存管理
+		CacheProviderHolder.init(config, (region, key)->{
+			//当一级缓存中的对象失效时，自动清除二级缓存中的数据
+			CacheProviderHolder.getLevel2Cache(region).evict(key);
+			log.debug(String.format("Level 1 cache object expired, evict level 2 cache object [%s,%s]", region, key));
+			if(policy != null)
+				policy.sendEvictCmd(region, key);
+		});
 
-			String cache_broadcast = props.getProperty("j2cache.broadcast");
-			if ("redis".equalsIgnoreCase(cache_broadcast)) {
-				String channel = props.getProperty("redis.channel");
-				policy = ClusterPolicyFactory.redis(channel, props);
-			}
-			else if ("jgroups".equalsIgnoreCase(cache_broadcast)) {
-				String channel_name = props.getProperty("jgroups.channel.name");
-				policy = ClusterPolicyFactory.jgroups(channel_name, props.getProperty("jgroups.configXml"), props);//
-			}
-			else
-				policy = ClusterPolicyFactory.custom(cache_broadcast, props);
-
-			log.info("Using cluster policy : " + policy.getClass().getName());
-		}
-	}
-
-	/**
-	 * get j2cache properties stream
-	 * @return
-	 */
-	private static InputStream getConfigStream() {
-		log.info("Load J2Cache Config File : [{}].", CONFIG_FILE);
-		InputStream configStream = J2Cache.class.getResourceAsStream(CONFIG_FILE);
-		if(configStream == null)
-			configStream = J2Cache.class.getClassLoader().getParent().getResourceAsStream(CONFIG_FILE);
-		if(configStream == null)
-			throw new CacheException("Cannot find " + CONFIG_FILE + " !!!");
-		return configStream;
+		policy = ClusterPolicyFactory.init(config.getBroadcast(), config.getBroadcastProperties());
+		log.info("Using cluster policy : " + policy.getClass().getName());
 	}
 
 }